SerialConnect: use ints for font metrics
As pointed by Ingo, using a float here is useless because we want each
line to be the same integer number of pixels. Now thigs are drawn
properly and the view has the exact same size as Terminal's one (tested
with font sizes 10 and 18, so different leading values don't seem to be
a problem).

Thanks for reviewing!

+++ src/apps/serialconnect/TermView.h
@@ -49,8 +49,8 @@ class TermView: public BView
 		VTerm* fTerm;
 		VTermScreen* fTermScreen;
 		BList fScrollBuffer;
-		float fFontWidth;
-		float fFontHeight;
+		int fFontWidth;
+		int fFontHeight;
 
 		static const VTermScreenCallbacks sScreenCallbacks;
